requirements.txt
pytest-playwright
playwright
allure-pytest
pytest
python-dotenv